+++ co2mpas/sampling/project.py
@@ -1570,11 +1570,11 @@ class OpenCmd(_SubCmd):
         %(cmd_chain)s [OPTIONS] <project>
     """
     def run(self, *args):
-        self.log.info('Opening project %r...', args)
         if len(args) != 1:
             raise CmdException(
                 "Cmd %r takes exactly one argument as the project-name, received %r!"
                 % (self.name, args))
+        self.log.info('Opening project %r...', args)
 
         projDB = self.projects_db
         proj = projDB.proj_open(args[0])
@@ -1738,7 +1738,9 @@ class ReportCmd(_SubCmd):
     - Eventually the *Dice Report* parameters will be time-stamped and disseminated to
       TA authorities & oversight bodies with an email, to receive back
       the sampling decision.
-    - To get report ready for sending it MANUALLY, use tstamp` sub-command.
+    - To send the report to the stamper, use `tsend` sub-command.
+    - To get report ready for sending it MANUALLY, use `tsend --dry-run` 
+      instead.
 
     """
 
@@ -1773,6 +1775,13 @@ class ReportCmd(_SubCmd):
 
 
 class TstampCmd(_SubCmd):
+    """Deprecated: renamed as `tsend`!"""
+    def run(self, *args):
+        raise CmdException("Cmd %r has been renamed to %r!"
+                               % (self.name, 'tsend'))
+
+
+class TsendCmd(_SubCmd):
     """
     IRREVOCABLY send report to the time-stamp service, or print it for sending it manually (--dry-run).
 
@@ -1947,6 +1956,7 @@ class TrecvCmd(TparseCmd):
     def run(self, *args):
         from . import tstamp
 
+        self.log.info("Receiving emails for projects(s) %s: ...", args)
         default_flow_style = None if self.verbose else False
         warn = self.log.warning
         rcver = tstamp.TstampReceiver(config=self.config)
@@ -1977,16 +1987,16 @@ class TrecvCmd(TparseCmd):
                      mid, pname)
                 continue
 
+            ## Respect verbose flag for print-outs.
+            infos = rcver._get_recved_email_infos(mail, verdict)
+            yield _mydump({mid: infos}, default_flow_style=default_flow_style)
+
             try:
                 proj.do_storedice(verdict=verdict)
                 #report = proj.result  # Not needed, we already have verdict.
             except Exception as ex:
                 self.log.error('Failed storing %s email, due to: %s',
                                mid, ex)
-
-            ## Respect verbose flag for print-outs.
-            infos2 = rcver._get_recved_email_infos(mail, verdict)
-            yield _mydump({mid: infos2}, default_flow_style=default_flow_style)
 
 
 class ExportCmd(_SubCmd):
@@ -2154,10 +2164,11 @@ class BackupCmd(_SubCmd):
     ).tag(config=True)
 
     def run(self, *args):
-        self.log.info('Archiving repo into %r...', args)
         if len(args) > 1:
             raise CmdException('Cmd %r takes one optional filepath, received %d: %r!'
                                % (self.name, len(args), args))
+        self.log.info('Archiving repo into %r...', args)
+        
         archive_fpath = args and args[0] or None
         kwds = {}
         if archive_fpath:
@@ -2178,6 +2189,8 @@ class BackupCmd(_SubCmd):
 
 all_subcmds = (LsCmd, InitCmd, OpenCmd,
                AppendCmd, ReportCmd,
-               TstampCmd, TrecvCmd, TparseCmd,
+               TstampCmd,  ## TODO: delete deprecated `projext tsend` cmd
+               TsendCmd,
+               TrecvCmd, TparseCmd,
                StatusCmd,
                ExportCmd, ImportCmd, BackupCmd)