+++ tests/sampling/test_tstamp.py
@@ -615,6 +615,39 @@ class TRX(unittest.TestCase):
             rcv.force = True
         self.check_timestamp(rcv, *verdicts)
 
+    @ddt.data(
+        (None, None, []),
+        (None, '', []),
+        (None, '  ', []),
+
+        ([], None, []),
+        ([], '', []),
+        ([], '  ', []),
+
+        (['', '  '], None, ['', '  ']),
+        (['', '  '], '', ['', '  ']),
+        (['', '  '], '  ', ['', '  ']),
+    )
+    def test_criteria_stripping(self, case):
+        ecrts, one, projects = case
+        rcv = tstamp.TstampReceiver(email_criteria=ecrts,
+                                    wait_criterio=one,
+                                    subject_prefix=one)
+        crt = rcv._prepare_search_criteria(True, projects)
+        self.assertEqual(crt, '')
+
+    def test_criteria_dupe_projects(self):
+        rcv = tstamp.TstampReceiver(email_criteria=[],
+                                    subject_prefix='')
+        crt = rcv._prepare_search_criteria(False, [])  # sanity
+        self.assertEqual(crt, '')
+
+        crt = rcv._prepare_search_criteria(False, ['ab', 'ab'])
+        self.assertNotIn('OR', crt)
+
+        crt = rcv._prepare_search_criteria(False, ['ab', 'foo', 'ab'])
+        self.assertEqual(crt.count('OR'), 1)
+
 
 @ddt.ddt
 class TstampShell(unittest.TestCase):